+++ java/com.sap.sailing.server.gateway.serialization/src/com/sap/sailing/server/gateway/deserialization/impl/ManeuverMainCurveWithEstimationDataJsonDeserializer.java
@@ -40,7 +40,7 @@ public class ManeuverMainCurveWithEstimationDataJsonDeserializer
                 .get(ManeuverMainCurveWithEstimationDataJsonSerializer.AVG_TURNING_RATE_IN_DEGREES_PER_SECOND);
         Double maxTurningRateInDegreesPerSecond = (Double) object
                 .get(ManeuverMainCurveWithEstimationDataJsonSerializer.MAX_TURNING_RATE_IN_DEGREES_PER_SECOND);
-        Long gpsFixesCount = (Long) object.get(ManeuverMainCurveWithEstimationDataJsonSerializer.GPS_FIXES_COUNT);
+        Integer gpsFixesCount = (Integer) object.get(ManeuverMainCurveWithEstimationDataJsonSerializer.GPS_FIXES_COUNT);
         return new ManeuverMainCurveWithEstimationDataImpl(boundaries.getTimePointBefore(),
                 boundaries.getTimePointAfter(), boundaries.getSpeedWithBearingBefore(),
                 boundaries.getSpeedWithBearingAfter(), boundaries.getDirectionChangeInDegrees(),
@@ -54,6 +54,6 @@ public class ManeuverMainCurveWithEstimationDataJsonDeserializer
                 boundaries.getDistanceSailedWithinManeuverTowardMiddleAngleProjection(),
                 boundaries.getDistanceSailedIfNotManeuvering(),
                 boundaries.getDistanceSailedTowardMiddleAngleProjectionIfNotManeuvering(),
-                avgTurningRateInDegreesPerSecond, gpsFixesCount.intValue());
+                avgTurningRateInDegreesPerSecond, gpsFixesCount);
     }
 }
